requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
loguru>=0.7.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
apscheduler>=3.10.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
playwright>=1.40.0
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
PyYAML>=6.0.0
lxml>=5.0.0
